moved flink error in flink package inside error folder and removed camel case import
