+++ internal/pkg/ccloudv2/flink_gateway_test.go
@@ -9,15 +9,16 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/confluentinc/cli/internal/pkg/errors"
+	"github.com/confluentinc/cli/internal/pkg/errors/flink"
 )
 
 func TestFlinkErrorCodeWhenErrors(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"errors":[{"detail":"There is an error"}]}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := errors.CatchFlinkError(errors.New("Some Error"), res)
+	err := flink.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(errors.FlinkError)
+	flinkError, ok := err.(flink.FlinkError)
 	require.True(t, ok)
 	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
@@ -26,15 +27,15 @@ func TestFlinkErrorCodeWhenErrors(t *testing.T) {
 func TestFlinkErrorNil(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"errors":[{"detail":"There is an error"}]}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := errors.CatchFlinkError(nil, res)
+	err := flink.CatchFlinkError(nil, res)
 	require.Nil(t, err)
 }
 
 func TestFlinkErrorNilHttpRes(t *testing.T) {
-	err := errors.CatchFlinkError(errors.New("Some Error"), nil)
+	err := flink.CatchFlinkError(errors.New("Some Error"), nil)
 	require.Error(t, err)
 
-	flinkError, ok := err.(errors.FlinkError)
+	flinkError, ok := err.(flink.FlinkError)
 	require.True(t, ok)
 	require.Equal(t, 0, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
@@ -43,10 +44,10 @@ func TestFlinkErrorNilHttpRes(t *testing.T) {
 func TestFlinkErrorCodeWhenErrorMessage(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"message":"unauthorized"}`)), StatusCode: http.StatusUnauthorized}
 
-	err := errors.CatchFlinkError(errors.New("Some Error"), res)
+	err := flink.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(errors.FlinkError)
+	flinkError, ok := err.(flink.FlinkError)
 	require.True(t, ok)
 	require.Equal(t, http.StatusUnauthorized, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
@@ -55,10 +56,10 @@ func TestFlinkErrorCodeWhenErrorMessage(t *testing.T) {
 func TestFlinkErrorCodeWhenNestedMessage(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader(`{"error":{"message":"gateway error"}}`)), StatusCode: http.StatusMethodNotAllowed}
 
-	err := errors.CatchFlinkError(errors.New("Some Error"), res)
+	err := flink.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(errors.FlinkError)
+	flinkError, ok := err.(flink.FlinkError)
 	require.True(t, ok)
 	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
@@ -67,10 +68,10 @@ func TestFlinkErrorCodeWhenNestedMessage(t *testing.T) {
 func TestFlinkErrorOnlyStatusCode(t *testing.T) {
 	res := &http.Response{Body: io.NopCloser(strings.NewReader("")), StatusCode: http.StatusMethodNotAllowed}
 
-	err := errors.CatchFlinkError(errors.New("Some Error"), res)
+	err := flink.CatchFlinkError(errors.New("Some Error"), res)
 	require.Error(t, err)
 
-	flinkError, ok := err.(errors.FlinkError)
+	flinkError, ok := err.(flink.FlinkError)
 	require.True(t, ok)
 	require.Equal(t, http.StatusMethodNotAllowed, flinkError.StatusCode())
 	require.Equal(t, err.Error(), flinkError.Error())
