+++ internal/pkg/flink/types/statement_error.go
@@ -3,7 +3,7 @@ package types
 import (
 	"fmt"
 
-	"github.com/confluentinc/cli/internal/pkg/errors"
+	"github.com/confluentinc/cli/internal/pkg/errors/flink"
 	"github.com/confluentinc/cli/internal/pkg/utils"
 )
 
@@ -53,7 +53,7 @@ func (e *StatementError) Error() string {
 
 // StatusCode extract the status code if the error implements Coder interface.
 func StatusCode(err error) int {
-	if coder, ok := err.(errors.Coder); ok {
+	if coder, ok := err.(flink.Coder); ok {
 		return coder.StatusCode()
 	}
 	return 0
